+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+- managed_object_storage: `managed_object_storage` resource public network validation
 
 ## [3.0.1] - 2023-10-23